Boston Globe-DOC /old code/requirements.txt
google-cloud-vision==2.0.0
python-dotenv=0.17.1
mysql-connector=2.2.9